+++ includes/functions-post-support_ticket.php
@@ -361,31 +361,3 @@ function set_default_ticket_meta( $post_id, $post, $update ) {
     }
 
 }
-
-
-function get_recent_tickets( $args = array() ) {
-
-    $defaults = array(
-        'author'  => '',
-        'after'   => 'now',
-        'before'  => '30 days ago',
-        'exclude' => array(),
-        'limit'   => -1
-    );
-
-    $args = wp_parse_args( $args, $defaults );
-
-
-    $q = array(
-        'post_type'      => 'support_ticket',
-        'post_status'    => 'publish',
-        'author'         => $args['author'],
-        'after'          => $args['after'],
-        'before'         => $args['before'],
-        'post__not_in'   => $args['exclude'],
-        'posts_per_page' => $args['limit'] ?: -1
-    );
-
-    return new \WP_Query( $q );
-
-}
